Move implementation of `rinq.ValidateNamespace` into new `nsutil` package to avoid cyclic dependency.
