+++ src/rinq/amqp/peer.go
@@ -10,6 +10,7 @@ import (
 	"github.com/rinq/rinq-go/src/rinq/internal/command"
 	"github.com/rinq/rinq-go/src/rinq/internal/localsession"
 	"github.com/rinq/rinq-go/src/rinq/internal/notify"
+	"github.com/rinq/rinq-go/src/rinq/internal/nsutil"
 	"github.com/rinq/rinq-go/src/rinq/internal/remotesession"
 	"github.com/rinq/rinq-go/src/rinq/internal/service"
 	"github.com/rinq/rinq-go/src/rinq/internal/syncutil"
@@ -105,7 +106,7 @@ func (p *peer) Session() rinq.Session {
 }
 
 func (p *peer) Listen(namespace string, handler rinq.CommandHandler) error {
-	if err := rinq.ValidateNamespace(namespace); err != nil {
+	if err := nsutil.Validate(namespace); err != nil {
 		return err
 	}
 
@@ -150,7 +151,7 @@ func (p *peer) Listen(namespace string, handler rinq.CommandHandler) error {
 }
 
 func (p *peer) Unlisten(namespace string) error {
-	if err := rinq.ValidateNamespace(namespace); err != nil {
+	if err := nsutil.Validate(namespace); err != nil {
 		return err
 	}
