+++ src/rinq/internal/nsutil/ginkgo_test.go
@@ -1,0 +1,13 @@
+package nsutil_test
+
+import (
+	"testing"
+
+	"github.com/onsi/ginkgo"
+	"github.com/onsi/gomega"
+)
+
+func TestSuite(t *testing.T) {
+	gomega.RegisterFailHandler(ginkgo.Fail)
+	ginkgo.RunSpecs(t, "nsutil")
+}
