+++ src/rinq/internal/remotesession/revision.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/rinq/rinq-go/src/rinq"
 	"github.com/rinq/rinq-go/src/rinq/ident"
+	"github.com/rinq/rinq-go/src/rinq/internal/nsutil"
 )
 
 type revision struct {
@@ -21,7 +22,7 @@ func (r *revision) Refresh(ctx context.Context) (rinq.Revision, error) {
 }
 
 func (r *revision) Get(ctx context.Context, ns, key string) (rinq.Attr, error) {
-	if err := rinq.ValidateNamespace(ns); err != nil {
+	if err := nsutil.Validate(ns); err != nil {
 		return rinq.Attr{}, err
 	}
 
@@ -40,7 +41,7 @@ func (r *revision) Get(ctx context.Context, ns, key string) (rinq.Attr, error) {
 }
 
 func (r *revision) GetMany(ctx context.Context, ns string, keys ...string) (rinq.AttrTable, error) {
-	if err := rinq.ValidateNamespace(ns); err != nil {
+	if err := nsutil.Validate(ns); err != nil {
 		return nil, err
 	}
 
@@ -70,7 +71,7 @@ func (r *revision) GetMany(ctx context.Context, ns string, keys ...string) (rinq
 }
 
 func (r *revision) Update(ctx context.Context, ns string, attrs ...rinq.Attr) (rinq.Revision, error) {
-	if err := rinq.ValidateNamespace(ns); err != nil {
+	if err := nsutil.Validate(ns); err != nil {
 		return nil, err
 	}
